# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Datespark API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "UserAdmin"
                ],
                "summary": "Allows admin to get list of users",
                "operationId": "9913e00853ba6b78c6d28fa18c81d214",
                "responses": {
                    "200": {
                        "description": "Returns list of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials or not authorized"
                    },
                    "422": {
                        "description": "Invalid body"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/{userId}": {
            "get": {
                "tags": [
                    "UserAdmin"
                ],
                "summary": "Allows admin to get user by id",
                "operationId": "470e800108cdd8ba30b0fc00d542f0c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user of given id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials or not authorized"
                    },
                    "422": {
                        "description": "Invalid body"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "put": {
                "tags": [
                    "UserAdmin"
                ],
                "summary": "Allows admin to update user",
                "operationId": "fa227bac9cbe44fabf2bcd5ae9af37fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials or not authorized"
                    },
                    "422": {
                        "description": "Invalid body"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "UserAdmin"
                ],
                "summary": "Allows admin to delete user",
                "operationId": "dd272c802ee2e3ff70c0bee9042c6c68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns if of deleted user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/DeletedModelResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials or not authorized"
                    },
                    "422": {
                        "description": "Invalid body"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DeletedModelResource": {
                "description": "Delete user",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "description": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jan Kowalski"
                    },
                    "surname": {
                        "type": "string",
                        "example": "jan@example.com"
                    },
                    "email": {
                        "type": "string",
                        "example": "jan@example.com"
                    },
                    "status": {
                        "description": "Either ACTIVE or DURING_REGISTER",
                        "type": "string",
                        "example": "ACTIVE"
                    },
                    "gender": {
                        "description": "Either MALE or FEMALE",
                        "type": "string",
                        "example": "MALE"
                    },
                    "roles": {
                        "description": "Array of roles of user in system",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "USER",
                                "ADMIN"
                            ]
                        },
                        "example": [
                            "USER"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}